requirements.txt
click
python-slugify
requests